requirements_minimal.txt
# Core Dependencies
google-generativeai>=0.8.5
requests>=2.32.0
python-dotenv>=1.1.0

# NLP & Tokenization
transformers>=4.40.0
tokenizers>=0.15.0

# Data Processing
pydantic>=2.0.0

# Utilities (optional but recommended)
tqdm>=4.67.0
